Maket everything parametric on price and volume
This also fixes a minor bug in `getOpenVol`.  This is a major overhaul.
I need to be able to have prices specified in ETH and BTC to be able to
trade in cross-crypto-currency markets, such as LTC-BTC or ETH-BTC. I
can now specify volumes in ETH and prices in BTC, for example.