Remove `cost` type variable
Including this was a mistake. `Cost` and `Price` are distinct
quantities, but the units of price (e.g. USD/BTC) and volume (e.g. BTC)
uniquely determine the unit of cost (e.g. (USD/BTC)*BTC = USD) So, we
don’t actually need another free type variable here. In the simple
model I am using, it is not possible to specify the cost in a different
unit. This just causes confusion.