Add client order id to Action type
The strategies need not know what the actual order ids are,
as long as they provide a unique client order id, the framework
is responsible for associating a real order id to the order to allow
for cancellation and other operations when the client order id is
provided. This commits enables this.