Make Ethereum have 8 d.p. - temp fix
I am getting numerical precision errors in ETH on the trading agent due
to using floating point calculations in a `Double`. For example,
comparisons that should be zero are differing by 1E-16. This commit
makes Ether look like other coins and avoid the problem, but it is a
temporary fix as Ethereum has 18 decimal places rather than 8...